Added HCFC statistics export
