+++ core/api/urls.py
@@ -25,7 +25,11 @@ from core.api.views.cp_archive import (
 )
 from core.api.views.cp_files import CPFilesDownloadView, CPFilesView
 from core.api.views.cp_prices import CPPricesView
-from core.api.views.cp_records_export import CPEmptyExportView, CPHFCExportView
+from core.api.views.cp_records_export import (
+    CPEmptyExportView,
+    CPHCFCExportView,
+    CPHFCExportView,
+)
 from core.api.views.cp_reports import (
     CPReportStatusUpdateView,
     CPReportView,
@@ -164,6 +168,11 @@ urlpatterns = [
         "country-programme/hfc/export/",
         CPHFCExportView.as_view(),
         name="country-programme-hfc-export",
+    ),
+    path(
+        "country-programme/hcfc/export/",
+        CPHCFCExportView.as_view(),
+        name="country-programme-hcfc-export",
     ),
     path(
         "country-programme/export-empty/",
